prettify code
